refactor(database): add some memory checkings into reading file
* Additionally, separate methods by files
